requirements.txt
appdirs==1.4.4
astroid==2.5
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
click==7.1.2
cycler==0.10.0
decorator==4.4.2
html5lib==1.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lxml==4.6.2
matplotlib==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
nltk==3.5
numpy==1.20.1
pandas==1.2.2
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pylint==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
regex==2020.11.13
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
soupsieve==2.2
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
